+++ Entity.h
@@ -12,7 +12,7 @@ namespace Entities
 
             virtual void input  (const sf::Event& e) = 0;
             virtual void input  () = 0;
-            virtual void update (float dt) = 0;
+            virtual void update (double dt) = 0;
             virtual void draw   () = 0;
     };
 }
